Create employee.py
in this exercise the program will print the address, request a pin number from the user, if the pin is correct the program will ask the user for a fruit to be checked. Next the program will search the list of fruits in the "sample.txt" file if the fruit is in the list it will tell the user so.